requirements.txt
django>=4.2.1
celery>=5.2.7
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
drf-spectacular>=0.26.2
django-ses>=3.5.0
psycopg2
psycopg2-binary
sentry-sdk>=1.24.0
wagtail==5.2.8
django_extensions>=3.2.1
django_storages>=1.13.2
drf-rw-serializers>=1.0.5
django-crum>=0.7.9
redis>=4.6.0
whitenoise>=6.4.0
django-phonenumber-field>=7.1.0
django[argon2]
django-phonenumber-field>=7.1.0
arrow>=1.2.3
phonenumbers>=8.13.13
